Win32:  Build fixes - change unsigned long long to uint64_t.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/branches/V5-4-patches@17608 06827809-a52a-0410-b366-d66718629ded

+++ net-snmp/include/net-snmp/data_access/interface.h
@@ -176,7 +176,7 @@ typedef struct netsnmp_interface_entry_s {
 typedef struct _conf_if_list {
     const char     *name;
     int             type;
-    unsigned long long speed;
+    uint64_t        speed;
     struct _conf_if_list *next;
 } netsnmp_conf_if_list;
